Add default empty embedded images
These images will normally be replaced by a proper
embedded image, but in some cases, it may be desirable
for a user to only load an image in one or a few embededed
CPUs.
If the user does not provide an image, this one is
used instead.
These images do nothing: at all effects that
CPU will be off.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>