Makefile: Support building C++ files with runner
Support providing C++ files to be built together with the runner, and
providing different options for C++ and C files, as well as selecting
the linker (by default NSI_CC = gcc).

Though this is intended only to build simple extensions.
When more complex addons are wanted, it would be recommended
to build them as a library which is then linked from the runner
instead of just lumping them into the runner build, as that way
there would be proper control of how those addons were built and no
effect on the runner build.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>